requirements.txt
asgiref==3.6.0
Django==4.1.5
django-solo==2.0.0
Pillow==9.4.0
python-dotenv==0.21.1
sqlparse==0.4.3
